+++ src/aux-server/aux-web/shared/interaction/DragOperation/BaseFileDragOperation.ts
@@ -168,8 +168,8 @@ export abstract class BaseFileDragOperation implements IOperation {
                 .transaction(
                     fileUpdated(this._file.id, {
                         tags: {
-                            'aux.mergeBall': null,
-                            'aux.mergeBall.tags': null,
+                            'aux.mod': null,
+                            'aux.mod.tags': null,
                         },
                     })
                 )