+++ CHANGELOG.md
@@ -9,9 +9,10 @@
 -   Improvements
     -   `onGridClick()` is now supported in VR.
     -   Changed `mergeBall` tag to `mod`.
+    -   Stop users from adding a blank or only whitespace tag.
 -   Bug Fixes
     -   All camera tweens will now snap to their final (and literal) target destination at the end of the tween.
-    -   Bots will get destroyed when dragged over the trashcan in AUX Builder even if still on a context surface.
+    -   Bots will get destroyed when dragged over the trashcan in AUX Builder even if it is still on a context surface.
 
 ## V0.9.4